+++ cfdm/mixin/__init__.py
@@ -13,6 +13,7 @@ from .parametersdomainancillaries import ParametersDomainAncillaries
 
 from .netcdf import (
     NetCDFComponents,
+    NetCDFConnectivityDimension,
     NetCDFGlobalAttributes,
     NetCDFGroupAttributes,
     NetCDFUnlimitedDimension,
@@ -22,6 +23,7 @@ from .netcdf import (
     NetCDFGroupsMixin,
     NetCDFChunks,
     NetCDFInterpolationSubareaDimension,
+    NetCDFMeshVariable,
     NetCDFMixin,
     NetCDFNodeCoordinateVariable,
     NetCDFSampleDimension,